[487] Rework tests
Duplicate the tests, one with feature flag enabled and one with it disabled. The test with the feature flag disabled can be deleted along with the feature flag.
